Update examples/genmap + function to dump map file
