+++ src/parrsb-aux.c
@@ -3,6 +3,7 @@
 
 #include <gencon.h>
 #include <genmap.h>
+#include <parRSB.h>
 
 #define MAXNV 8  /* maximum number of vertices per element */
 #define MAXDIM 3 /* maximum number of vertices per element */
@@ -18,48 +19,6 @@ typedef struct {
   long long vtx[MAXNV];
   double coord[MAXNV * MAXDIM];
 } elem_data;
-
-int parrsb_read_mesh(unsigned int *nel, int *nv, long long **vl, double **coord,
-                     char *name, MPI_Comm comm, int read) {
-  struct comm c;
-  comm_init(&c, comm);
-
-  Mesh mesh;
-
-  /* Read geometry from .re2 file */
-  if (read & 1) {
-    char geom_name[BUFSIZ];
-    strncpy(geom_name, name, BUFSIZ);
-    strncat(geom_name, ".re2", 5);
-
-    read_geometry(&mesh, geom_name, &c);
-    get_vertex_coordinates(coord, mesh);
-  }
-
-  /* Read connectivity from .co2 file */
-  if (read & 2) {
-    char conn_name[BUFSIZ];
-    strncpy(conn_name, name, BUFSIZ);
-    strncat(conn_name, ".co2", 5);
-
-    assert(read & 1);
-    read_connectivity(mesh, conn_name, &c);
-    get_vertex_ids(vl, mesh);
-  }
-
-  *nel = (unsigned int)get_mesh_nel(mesh);
-
-  int ndim = get_mesh_dim(mesh);
-#if defined(MPI)
-  MPI_Bcast(&ndim, 1, MPI_INT, 0, MPI_COMM_WORLD);
-#endif
-  *nv = (ndim == 3) ? 8 : 4;
-
-  mesh_free(mesh);
-  comm_free(&c);
-
-  return 0;
-}
 
 int parrsb_distribute_elements(unsigned int *nelt_, long long **vl_,
                                double **coord_, int *part, int nv,
@@ -129,86 +88,6 @@ int parrsb_distribute_elements(unsigned int *nelt_, long long **vl_,
 
   return 0;
 }
-
-#define WRITE_INT(dest, val)                                                   \
-  do {                                                                         \
-    memcpy(dest, &(val), sizeof(int));                                         \
-  } while (0)
-
-int parrsb_dump_con(long long *vl, unsigned int nelt, int nv, char *name,
-                    MPI_Comm comm) {
-  const char version[5] = "#v001";
-  const float test = 6.54321;
-
-  struct comm c;
-  comm_init(&c, comm);
-
-  uint id = c.id;
-
-  MPI_File file;
-  int err = MPI_File_open(comm, name, MPI_MODE_CREATE | MPI_MODE_WRONLY,
-                          MPI_INFO_NULL, &file);
-  if (err) {
-    if (id == 0) {
-      fprintf(stderr, "%s:%d Error opening file: %s for writing.\n", __FILE__,
-              __LINE__, name);
-      fflush(stdout);
-    }
-    return err;
-  }
-
-  slong out[2][1], bfr[2][1];
-  slong in = nelt;
-  comm_scan(out, &c, gs_long, gs_add, &in, 1, bfr);
-  slong start = out[0][0];
-  slong nelgt = out[1][0];
-  slong nelgv = nelgt;
-
-  int write_size = nelt * (nv + 1) * sizeof(int);
-  int header_size = GC_CO2_HEADER_LEN + sizeof(float);
-  if (id == 0)
-    write_size += header_size;
-
-  char *buf = (char *)calloc(write_size, sizeof(char));
-  char *buf0 = buf;
-
-  if (id == 0) {
-    sprintf(buf0, "%5s%12d%12d%12d", version, (int)nelgt, (int)nelgv, nv);
-#if 0
-    printf("%5s%12d%12d%12d\n", version, (int)nelgt, (int)nelgv, nv);
-#endif
-    memset(buf0 + strlen(buf0), ' ', GC_CO2_HEADER_LEN - strlen(buf0));
-    buf0[GC_CO2_HEADER_LEN] = '\0';
-    buf0 += GC_CO2_HEADER_LEN;
-    memcpy(buf0, &test, sizeof(float));
-    buf0 += sizeof(float);
-  }
-
-  int i, j, temp;
-  for (i = 0; i < nelt; i++) {
-    temp = start + i + 1;
-    WRITE_INT(buf0, temp);
-    buf0 += sizeof(int);
-    for (j = 0; j < nv; j++) {
-      temp = vl[i * nv + j];
-      WRITE_INT(buf0, temp);
-      buf0 += sizeof(int);
-    }
-  }
-
-  MPI_Status st;
-  err |= MPI_File_write_ordered(file, buf, write_size, MPI_BYTE, &st);
-  err |= MPI_File_close(&file);
-
-  comm_barrier(&c);
-  comm_free(&c);
-
-  free(buf);
-
-  return err;
-}
-
-#undef WRITE_INT
 
 void parrsb_part_stat(long long *vtx, int nel, int nv, MPI_Comm ce) {
   int i, j;
